+++ doc/basics.md
@@ -180,6 +180,32 @@ behavior.
 > parser.parse(json).tie(doc, error); // <-- Assigns to doc and error just like "auto [doc, error]"
 > ```
 
+
+We can write a "quick start" example where we attempt to parse a file and access some data, without triggering exceptions:
+
+```C++
+#include "simdjson.h"
+
+int main(void) {
+  simdjson::dom::parser parser;
+  simdjson::dom::element tweets;
+  simdjson::error_code error;
+  parser.load("twitter.json").tie(tweets,error);
+  if(error) {
+    std::cerr << "could not load document" << std::endl;
+    return EXIT_FAILURE;
+  }
+  simdjson::dom::element res;
+
+  tweets["search_metadata"]["count"].tie(res,error);
+  if(error) {
+    std::cerr << "could not access keys" << std::endl;
+    return EXIT_FAILURE;
+  }
+  std::cout << res << " results." << std::endl;
+}
+```
+
 ### Error Handling Example
 
 This is how the example in "Using the Parsed JSON" could be written using only error code checking:
